requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
bech32==1.2.0
betterproto==2.0.0b4
bip32utils==0.3.post4
boltons==21.0.0
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
configparser==5.2.0
ecdsa==0.17.0
frozenlist==1.3.0
furl==2.1.3
grpclib==0.4.2
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
mnemonic==0.19
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.22.4
opencv-python==4.5.5.64
orderedmultidict==1.0.1
protobuf==3.20.1
pure-python-adb==0.3.0.dev0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
terra-proto==2.0.0
terra-sdk==3.0.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
wrapt==1.14.1
yarl==1.7.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5